+++ trans.go
@@ -2,5 +2,5 @@ package timeago
 
 type Translation struct {
 	Language     string
-	Translations LangSet
+	Translations *LangSet
 }