+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # 🗒 Release Notes
 
 ## v3.0.0 (2024-11-10)
-- Changed package namespace to `github.com/SerhiiCho/timeago/v3` to match Go versioning rules
+- Breaking changes:
+    - Changed package namespace to `github.com/SerhiiCho/timeago/v3`
+    - Rename `SetConfig` function to `Configure`
+    - Moved `timeago.Config` object to `config.Config` package
+    - `Configure` function now excepts pointer to the config `timeago.Configure(&config.Config{})` instead of just `config.Config{}`
 
 ## v2.2.1 (2024-03-28)
 - Updated `README.md` file by adding more information about supported languages